#42: move everything to customer.online.portal package
