--- customer-online-portal/src/main/java/com/rppjs/customer/Application.java
@@ -1,12 +1,0 @@
-package com.rppjs.customer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
