--- customer-online-portal/src/main/java/com/rppjs/customer/dtos/BaseResponseDTO.java
@@ -1,7 +1,0 @@
-package com.rppjs.customer.dtos;
-
-public class BaseResponseDTO {
-
-    protected String message;
-    protected Exception e;
-}
