--- customer-online-portal/src/main/java/com/rppjs/customer/dtos/RegistrationRequestDTO.java
@@ -1,9 +1,0 @@
-package com.rppjs.customer.dtos;
-
-public class RegistrationRequestDTO {
-
-    public String email;
-    public String pass;
-    public String firstName;
-    public String lastName;
-}
