--- customer-online-portal/src/main/java/com/rppjs/customer/dtos/RegistrationResponseDTO.java
@@ -1,6 +1,0 @@
-package com.rppjs.customer.dtos;
-
-public class RegistrationResponseDTO {
-
-    public String email;
-}
