+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/Application.java
@@ -1,0 +1,12 @@
+package com.rppjs.customer.online.portal;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
