+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/dtos/BaseResponseDTO.java
@@ -1,0 +1,7 @@
+package com.rppjs.customer.online.portal.dtos;
+
+public class BaseResponseDTO {
+
+    protected String message;
+    protected Exception e;
+}
