+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/dtos/RegistrationRequestDTO.java
@@ -1,0 +1,9 @@
+package com.rppjs.customer.online.portal.dtos;
+
+public class RegistrationRequestDTO {
+
+    public String email;
+    public String pass;
+    public String firstName;
+    public String lastName;
+}
