+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/dtos/RegistrationResponseDTO.java
@@ -1,0 +1,6 @@
+package com.rppjs.customer.online.portal.dtos;
+
+public class RegistrationResponseDTO {
+
+    public String email;
+}
