+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/repository/CustomerRepository.java
@@ -1,0 +1,10 @@
+package com.rppjs.customer.online.portal.repository;
+
+import com.rppjs.customer.online.portal.entities.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
+}
