+++ customer-online-portal/src/main/java/com/rppjs/customer/online/portal/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.rppjs.customer.online.portal.repository;
+
+import com.rppjs.customer.online.portal.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+
+}
