--- customer-online-portal/src/main/java/com/rppjs/customer/repository/CustomerRepository.java
@@ -1,10 +1,0 @@
-package com.rppjs.customer.repository;
-
-import com.rppjs.customer.entities.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-
-}
